test(protocol): extract blame generation tests
